Restructure app for Next.js
