+++ components/BarChart.js
