+++ components/HotspotsList.js
