+++ components/SearchBar.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { withRouter } from 'react-router-dom'
+import { withRouter } from 'next/router'
 import { Input } from 'antd'
 import Client from '@helium/http'
