+++ components/TxnReward.js
