+++ next.config.js
@@ -1,0 +1,2 @@
+const withImages = require('next-images')
+module.exports = withImages()
