+++ package.json
@@ -15,6 +15,9 @@
     "export-to-csv": "^0.2.1",
     "h3-js": "3.6.4",
     "mapbox-gl": "1.10.1",
+    "next": "^9.5.5",
+    "next-ga": "^2.3.4",
+    "next-images": "^1.6.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-grid-system": "^6.3.1",
@@ -30,10 +33,9 @@
     "recharts": "^1.8.5"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "pretty": "prettier --write \"./**/*.{js,jsx,json,css}\""
   },
   "husky": {
