+++ pages/TxnView.js
