+++ pages/_app.js
@@ -1,0 +1,23 @@
+import Router from 'next/router'
+import withGA from 'next-ga'
+
+import '../styles/index.css'
+import '../styles/Explorer.css'
+
+function MyApp({ Component, pageProps }) {
+  return <Component {...pageProps} />
+}
+
+// Only uncomment this method if you have blocking data requirements for
+// every single page in your application. This disables the ability to
+// perform automatic static optimization, causing every page in your app to
+// be server-side rendered.
+//
+// MyApp.getInitialProps = async (appContext) => {
+//   // calls page's `getInitialProps` and fills `appProps.pageProps`
+//   const appProps = await App.getInitialProps(appContext);
+//
+//   return { ...appProps }
+// }
+
+export default withGA('UA-52432858-12', Router)(MyApp)
