+++ pages/hotspots/[hotspotid].js
@@ -3,8 +3,8 @@ import { Row, Typography, Checkbox, Tooltip, Card, Table } from 'antd'
 import Client from '@helium/http'
 import round from 'lodash/round'
 import get from 'lodash/get'
-import AppLayout, { Content } from '../components/AppLayout'
-import ActivityList from '../components/ActivityList'
+import AppLayout, { Content } from '../../components/AppLayout'
+import ActivityList from '../../components/ActivityList'
 import ReactMapboxGl, { Layer, Marker, Feature } from 'react-mapbox-gl'
 import Fade from 'react-reveal/Fade'
 import HotspotImg from '../images/hotspot.svg'
