+++ styles/hotspots.json
